Cast dma semaphore payload to ulong to fill the space

+++ Ryujinx.Graphics.Gpu/Engine/Dma/DmaClass.cs
@@ -115,7 +115,7 @@ namespace Ryujinx.Graphics.Gpu.Engine.Dma
                 }
                 else /* if (type == LaunchDmaSemaphoreType.ReleaseFourWordSemaphore) */
                 {
-                    _channel.MemoryManager.Write(address, _state.State.SetSemaphorePayload);
+                    _channel.MemoryManager.Write(address, (ulong)_state.State.SetSemaphorePayload);
                     _channel.MemoryManager.Write(address + 8, _context.GetTimestamp());
                 }
             }
